backend/requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy
psycopg2-binary
python-dotenv
pydantic-settings
azure-storage-blob
azure-storage-queue
aiohttp
python-multipart